Added Graphs + Data Splits
I added the graphs after the 'Gender Unification" portion of the code. This is done before we transform the code to numeric in order to keep the factor variable labels. 

Towards the end of the script I added code that resamples the code (mixes it up) and then splits it into test/ train sets.

I am about to get something to eat and when I get back I will work on it some more.